refactor: Separate Footer.module.scss from Home.module.scss
