+++ components/footer/Footer.module.scss
@@ -1,0 +1,27 @@
+@import '@/styles/variables';
+
+.footer {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  height: 100px;
+  border-top: 1px solid $light-gray;
+
+  span {
+    display: flex;
+  }
+
+  a {
+    display: flex;
+    flex-grow: 1;
+    align-items: center;
+    justify-content: center;
+    color: inherit;
+    text-decoration: none;
+  }
+}
+
+.twitterAccount {
+  margin: 0 4px;
+}
