+++ pages/index.tsx
@@ -5,8 +5,8 @@ import timezone from 'dayjs/plugin/timezone';
 import utc from 'dayjs/plugin/utc';
 import { GetStaticProps } from 'next';
 import Head from 'next/head';
-import Footer from '@/components/Footer';
 import Card from '@/components/card/Card';
+import Footer from '@/components/footer/Footer';
 import Header from '@/components/header/Header';
 import { Pagination } from '@/components/pagination/Pagination';
 import { microcms } from '@/libs/microcms';
