+++ styles/Home.module.scss
@@ -19,28 +19,6 @@
   padding: 80px 0;
 }
 
-.footer {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  height: 100px;
-  border-top: 1px solid $light-gray;
-
-  span {
-    display: flex;
-  }
-
-  a {
-    display: flex;
-    flex-grow: 1;
-    align-items: center;
-    justify-content: center;
-    color: inherit;
-    text-decoration: none;
-  }
-}
-
 .grid {
   display: flex;
   flex-wrap: wrap;
@@ -49,27 +27,9 @@
   width: 80%;
 }
 
-.twitterAccount {
-  margin: 0 4px;
-}
-
 @media (max-width: 600px) {
   .grid {
     flex-direction: column;
     width: 95%;
-  }
-
-  .title {
-    font-size: 2rem;
-  }
-
-  .description {
-    font-size: 1rem;
-  }
-
-  .card {
-    h2 {
-      font-size: 1rem;
-    }
   }
 }
